requirements.txt
openai>=1.3.0,<2.0
langchain>=0.1.0,<0.2.0
spotipy>=2.23.0,<3.0
python-dotenv>=1.0.0,<2.0
fastapi>=0.100,<0.110
uvicorn[standard]>=0.22.0,<0.25
pydantic>=2.0,<3.0
mypy>=1.0,<2.0
